app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
